+++ src/components/MapComponents/BoothCard.jsx
@@ -1,30 +1,29 @@
 import React, { useState } from "react";
 import HeartIcon from "../../assets/images/icons/map-icons/Heart.svg";
 import UnheartIcon from "../../assets/images/icons/map-icons/Unheart.svg";
-import PictureIcon from "../../assets/images/icons/map-icons/Picture.png";
+import Badge from "./Badge";
 
 const BoothCard = ({
   title,
-  location,
-  time,
   image,
-  isOperating = false,
-  className = "",
-  likeCount = 12,
-  badges = { isEventActive: false, isDOrderPartner: false },
+  isNight,
+  startTime,
+  endTime,
+  businessDays,
+  location,
+  isOperating,
+
+  likesCount = 0,
+  isLiked,
+  isEvent,
+  isDorder,
 }) => {
-  const [isLiked, setIsLiked] = useState(false);
-
-  const handleHeartClick = (e) => {
-    e.stopPropagation();
-    setIsLiked(!isLiked);
-  };
-
   return (
     <div
-      className={`bg-white rounded-2xl border p-4 ${
-        isOperating ? "border-primary-400" : "border-neutral-200"
-      } ${className}`}
+      className={`bg-white w-full
+        rounded-2xl border p-4 ${
+          isOperating ? "border-primary-400" : "border-neutral-200"
+        } `}
       style={{
         boxShadow: "0 3px 5px 0 rgba(0, 0, 0, 0.10)",
       }}
@@ -33,30 +32,16 @@ const BoothCard = ({
         <div className="relative w-16 h-16 flex-shrink-0">
           <div className="w-16 h-16 bg-gray-100 rounded-lg overflow-hidden">
             <img
-              src={PictureIcon}
+              src={image}
               alt="부스 이미지"
               className="w-full h-full object-cover"
             />
           </div>
-
-          {badges.isEventActive && (
-            <div className="absolute -top-3 -left-1">
-              <span
-                className="inline-flex px-2 py-0.5 justify-center items-center rounded-xl text-white text-[10px] font-normal leading-[150%] font-suite shadow-tag"
-                style={{ background: "rgba(239, 112, 99, 0.90)" }}
-              >
-                Event
-              </span>
-            </div>
-          )}
         </div>
 
         <div className="flex-1 relative">
           <div className="absolute top-0 right-0 flex flex-col items-center">
-            <button
-              onClick={handleHeartClick}
-              className="w-6 h-6 flex items-center justify-center mb-1"
-            >
+            <button className="w-6 h-6 flex items-center justify-center mb-1">
               <img
                 src={isLiked ? HeartIcon : UnheartIcon}
                 alt="좋아요"
@@ -64,12 +49,12 @@ const BoothCard = ({
               />
             </button>
             <span className="text-xs text-neutral-300 font-suite">
-              {likeCount}
+              {likesCount}
             </span>
           </div>
 
           <p className="text-xs text-neutral-400 mb-0.5 font-suite leading-[150%] font-normal">
-            {time}
+            {businessDays} {startTime}~{endTime}
           </p>
 
           <h3 className="text-xl font-semibold text-black mb-0.5 font-suite leading-[130%]">
@@ -80,15 +65,6 @@ const BoothCard = ({
             <p className="text-sm text-black font-suite leading-[150%] font-normal">
               {location}
             </p>
-
-            {badges.isDOrderPartner && (
-              <span
-                className="inline-flex px-2 py-0.5 justify-center items-center rounded-xl text-white text-[10px] font-normal leading-[150%] font-suite shadow-tag"
-                style={{ background: "rgba(248, 176, 169, 0.90)" }}
-              >
-                D-order
-              </span>
-            )}
           </div>
         </div>
       </div>
